separated  userView
